fix : searchController -> ClubController 로 변경 (#773)
